Completely revamped Player.java and project structure
BasicBot contains universal bot code, it is extended by other bots, such
as WorkerBot.

Player.java now takes a census each round of all the bots and sorts them
into lists, which are passed to the appropriate Manager, ie a List of
WorkerBots is sent to WorkerManager.

All Managers have a list of common methods they need to implement
outlined in the interface, specific implentation is up to the Manager.

The test case written involved sending the starting workers to an
arbitrary point, and then later on to another one.  The sequence of
events is as follows:
1) round 1 the Player makes a list of WorkerBots and gives it to the
WorkerManager.
2)The Player does workerManager.issueOrderMoveAllUnits(MapLocation)
3)The WorkerManager object does all appropriate method, which in this
case assigns all the units the target destination
4)The Player tells the workerManager to go ahead and
eachTurnMoveAllUnits(), which the Manager dutifully does, moveing the
workers to their targetDestination.