initial Ranger orders and action tree
Works similarly to last year's bots. Each bot has a given task and a location for that task. It attempts to move to that location and perform its task on the way.

the attack mode is standard attack move to a lcoation.

protect mode will go to a location, attempt to find a unit. if it finds a unit, it orbits the unit, attacking enemies as it goes.

The move mode goes to the destination without attacking.

todo:snipe, idle, and wait modes, actually populating the modes