Added Order Stack to all units (untested)
Use the Stack to assign orders to individual units.  An order consists
of a location and an enum.  This allows bots to remember previous tasks
and to resume them if they are interrupted with new orders.

Moved the nuts and bolts type functionality from WorkerManager to
WorkerBot.  WorkerManager now just assigns orders to the bots, and bots
are able to complete them on their own and then move to the next one.

Order of events for a worker traveling to a resource deposit and passing
by a blueprint -
1) Worker has one order in its stack, Harvest at Location A.
2) Navigates near an unfinished blueprint.
3)WorkerManager sees this and pushes a new order onto the stack : build
blueprint at location B.
4) since the build blueprint order is on top of the stack, the robot
navigates to that location.  Once it is unable to build, either due to
it being done, non-existant, or on the wrong team (just in case... don't
expect it to happen) , it'll pop that order and continue on toward
location A to harvest