Enabled worker replication
Added player controlled variable for target worker population, and can
customize it based on planet.

Added function for player.java to add new units to the basicRobotMap.
What was happening with new WorkerBots is that the vecUnit data
structure was giving us all the workers, our code would grab the
BasicBot from the map, and since we didn't have anything mapped to that
ID yet it would grab null.... which of course caused all kinds of issues
very quickly.